refactor: Improve Car.move() for better extensibility
