+++ src/domain/validate.js
