+++ src/index.js
@@ -1,4 +1,4 @@
-import { startRace } from "./Race.js";
+import { startRace } from "./domain/Race.js";
 
 async function run() {
   await startRace();
